+++ calc.h
@@ -1,0 +1,10 @@
+#ifndef CALC_H_INCLUDED
+#define CALC_H_INCLUDED
+
+void add(double x, double y);
+void subtract(double x, double y);
+void multiply(double x, double y);
+void divide(double x, double y);
+void calc();
+
+#endif // CALC_H_INCLUDED