+++ help.cpp
@@ -6,5 +6,7 @@ void help_generic() {
     std::cout << "version   - Shows the NShell build number.\n";
     std::cout << "c.version - Same as version, but in a single line.\n";
     std::cout << "contrib   - Lists all contributors of NShell.\n";
+    std::cout << "license   - Shows the license of NShell.\n";
+    std::cout << "calc      - A basic calculator included in the NShell.\n";
     std::cout << "exit      - Closes the NShell.\n";
 }