+++ main.cpp
@@ -1,7 +1,31 @@
 #include "help.h"
+#include "calc.h"
 #include <iostream>
 
-int buildNr = 6;
+int buildNr = 7;
+
+const char *mitLicense =
+"MIT License\n\n"
+
+"Copyright (c) 2023 Nicolas Lucien and NShell contributors\n\n"
+
+"Permission is hereby granted, free of charge, to any person obtaining a copy "
+"of this software and associated documentation files (the \"Software\"), to deal "
+"in the Software without restriction, including without limitation the rights "
+"to use, copy, modify, merge, publish, distribute, sublicense, and/or sell "
+"copies of the Software, and to permit persons to whom the Software is "
+"furnished to do so, subject to the following conditions:\n\n"
+
+"The above copyright notice and this permission notice shall be included in all "
+"copies or substantial portions of the Software.\n\n"
+
+"THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR "
+"IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, "
+"FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE "
+"AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER "
+"LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, "
+"OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE "
+"SOFTWARE.\n";
 
 int main()
 {
@@ -30,12 +54,23 @@ int main()
             help_generic();
         }
 
+        else if (command_to_run == "calc") {
+            calc();
+        }
+
+        else if (command_to_run == "license") {
+            std::cout << "NShell\n" << "Build " << buildNr << '\n';
+            std::cout << "\n";
+
+            std::cout << mitLicense;
+        }
+
         else if (command_to_run == "exit") {
             break;
         }
 
         else {
-            std::cout << "Unknown command!\n";
+            std::cout << "NShell encountered an error: Unknown command!\n";
         }
     }